+++ tests/test_google_ads_interrupted_sync_full_table.py
@@ -36,6 +36,7 @@ class InterruptedSyncFullTableTest(GoogleAdsBase):
         streams_under_test = {
             'ads',
             'campaign_criterion',
+            'assets'
         }
     
         # Create connection