+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v2.0.0
   * Updates API version to 20
+  * Add new stream `assets`
+  * Remove feed-related streams
   * Updates pkg version to 27.0.0
   * Updates `requests` to 2.32.4
   * Updates `singer-python` to 6.1.1